some more component cleanup
