+++ lib/Components/Node.js
@@ -1,0 +1,11 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.Node = Node;
+var React = require("react");
+function Node(_a) {
+    var renderNodeWrapper = _a.renderNodeWrapper, index = _a.index, children = _a.children;
+    if (renderNodeWrapper) {
+        return React.createElement(React.Fragment, null, renderNodeWrapper(children, index));
+    }
+    return React.createElement(React.Fragment, null, children);
+}
