+++ lib/Components/Title.d.ts
@@ -1,0 +1,3 @@
+import * as React from "react";
+import { ReactDiffViewerRenderProps } from "../getLinesToRender";
+export declare function Title({ hideLineNumbers, leftTitle, rightTitle, splitView, styles, }: ReactDiffViewerRenderProps): React.JSX.Element;
