+++ lib/helpers.js
@@ -25,6 +25,7 @@ var __spreadArray = (this && this.__spreadArray) || function (to, from, pack) {
     return to.concat(ar || Array.prototype.slice.call(from));
 };
 Object.defineProperty(exports, "__esModule", { value: true });
+exports.expandBlock = expandBlock;
 function resetCodeBlocks(expandedBlocksSize, setExpandedBlocks) {
     if (expandedBlocksSize > 0) {
         setExpandedBlocks(new Set());
